TASK: Mark `InMemoryEventStore` as experimental api and move to `Adapter`
